content/blog/2024-scientific-publishing/requirements.txt
aiodns==3.2.0
aiofiles==22.1.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
alpha-vantage==2.3.1
altair==5.3.0
ansiwrap==0.8.4
anyio==4.3.0
appdirs==1.4.4
appnope==0.1.4
arch==5.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
ascii-magic==1.6
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
backoff==2.2.1
base58==2.1.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
bs4==0.0.1
bt==0.2.9
cachetools==5.3.3
ccxt==4.3.18
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
commonmark==0.9.1
contourpy==1.2.1
convertdate==2.4.0
cryptography==42.0.7
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.12.1
Cython==3.0.10
dataclasses-json==0.5.14
dateparser==1.2.0
DateTime==5.5
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
degiro-connector==2.0.21
detecta==0.0.5
distro==1.9.0
dnspython==2.6.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.1
exchange_calendars==4.5.4
executing==2.0.1
fastjsonschema==2.19.1
feedparser==6.0.11
ffn==0.3.6
financedatabase==2.1.0
finnhub-python==2.4.19
finviz==1.4.6
finvizfinance==0.14.7
flake8==7.0.0
fonttools==4.51.0
formulaic==1.0.1
fqdn==1.5.1
fred==3.1
fredapi==0.4.3
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.3.1
fundamentalanalysis==0.2.14
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
hijri-converter==2.3.1
holidays==0.14.2
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
interface-meta==1.3.0
intrinio-sdk==6.27.2
ipykernel==6.29.4
ipympl==0.8.4
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
iso8601==0.1.16
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_fileid==0.9.2
jupyter_server_terminals==0.5.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
langchain==0.0.237
langsmith==0.0.10
linearmodels==4.31
llama-index==0.7.11
loguru==0.7.2
lxml==5.2.1
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.0.2
monotonic==1.6
more-itertools==10.2.0
mplfinance==0.12.10b0
mstarpy==0.0.4
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==6.5.7
notebook_shim==0.2.4
numexpr==2.10.0
numpy==1.23.4
oandapyV20==0.6.3
onetimepass==1.0.1
openai==1.26.0
openapi-schema-pydantic==1.2.4
openbb==3.2.0
openpyxl==3.1.2
overrides==7.7.0
packaging==24.0
pandas==1.5.3
pandas-datareader==0.10.0
pandas-market-calendars==3.2
pandas_ta==0.3.14b0
pandocfilters==1.5.1
papermill==2.4.0
parso==0.8.4
patsy==0.5.6
peewee==3.17.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
pmaw==3.0.0
posthog==3.5.0
praw==7.7.1
prawcore==2.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
property-cached==1.6.4
protobuf==3.20.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pycares==4.4.0
pycodestyle==2.11.1
pycoingecko==3.1.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.15
pydeck==0.9.0
pyflakes==3.2.0
Pygments==2.18.0
pyhdfe==0.2.0
pyluach==2.2.0
PyMeeus==0.5.12
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyotp==2.9.0
pyparsing==3.1.2
PyPrind==2.11.3
pyrsistent==0.19.3
pythclient==0.1.24
python-binance==1.0.19
python-coinmarketcap==0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-i18n==0.3.9
python-json-logger==2.0.7
pytrends==4.9.2
pytz==2024.1
pywry==0.5.14
PyYAML==6.0.1
pyzmq==26.0.3
Quandl==3.7.0
rdflib==7.0.0
referencing==0.35.1
regex==2024.4.28
reportlab==3.6.13
requests==2.31.0
requests-futures==1.0.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
robin-stocks==2.1.0
rpds-py==0.18.1
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.2
scipy==1.10.1
screeninfo==0.6.7
seaborn==0.11.2
Send2Trash==1.8.3
setproctitle==1.3.3
setuptools-scm==8.1.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.30
squarify==0.4.3
stack-data==0.6.3
statsmodels==0.13.5
stocksera==0.1.21
streamlit==1.34.0
svglib==1.5.1
tabula-py==2.9.0
tabulate==0.9.0
tenacity==8.3.0
terminado==0.18.1
textwrap3==0.9.2
thepassiveinvestor==1.2.1
threadpoolctl==3.5.0
tiktoken==0.6.0
tinycss2==1.3.0
tokenterminal==1.0.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
tradingview-ta==3.3.0
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
update-checker==0.18.0
uri-template==1.3.0
urllib3==1.26.18
user_agent==0.1.10
vaderSentiment==3.3.2
valinvest==0.0.2
watchdog==2.3.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
widgetsnbextension==4.0.10
wrapt==1.16.0
y-py==0.6.2
yahooquery==2.3.2
yarl==1.9.4
yfinance==0.2.38
ypy-websocket==0.8.4
zope.interface==6.3
